Refactor test servers for Redis and Spanner
