+++ lib/srv/db/spanner/testing/test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package spanner
+package testing
 
 import (
 	"context"
